pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "logforge"
version = "1.2.0"
description = "logforge notification system"
authors = ["LogForge <logforge@gmail.com>"]
readme = "README.md"

packages = [
    { include = "logforge" }
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pyyaml = ">=6.0"
requests = ">=2.28.0"
meta-ai-api = ">=1.2.5"
uvicorn = "^0.34.2"
fastapi = "^0.115.12"
apprise = "^1.9.3"

[tool.poetry.scripts]
LogForge = "logforge.server:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pre-commit = "^4.2.0"
deptry = "^0.23.0"

[tool.black]
target-version = ["py311"]
preview = true
line-length = 120

[tool.ruff]
target-version = "py311"
line-length = 120
include = [
    "logforge/**/*.py",
    "*.py"
]
